requirements.txt
playwright>=1.48.0
httpx>=0.25.0
deepdiff>=6.7.0
pyyaml>=6.0.1
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
google-generativeai>=0.3.0

